requirements.txt
fastapi
# uvicorn[standard]
uvicorn
openai
gunicorn
python-dotenv
asyncpg
google-cloud-bigquery